Join two charts into one as Chart View to handle both cuantitative and cualitaitve data.
