+++ medTracker/View/Analysis/Chart/CualitativeGraph/BarChartView_Cual.swift
